Don't keep objects themselves in indirect type
Now the type indirect doesn't hold the objects to be written to the
output. Instead, PDF objects are instantly written to the output as soon
as they're finished. This also avoids keeping everything in the memory.

I also did some cleanup in this commit: comments starting with the
capital, renaming some vars, improving error-handling, etc.